+++ CSFML/src/Graphics/RenderWindow.cpp
@@ -212,3 +212,17 @@ extern "C" void sfRenderWindow_popGLStates(sf::RenderWindow *renderWindow) {
 extern "C" void sfRenderWindow_resetGLStates(sf::RenderWindow *renderWindow) {
     renderWindow->resetGLStates();
 }
+
+extern "C" void sfMouse_setPositionRenderWindow(sfVector2i pos, const sf::RenderWindow *relativeTo) {
+    sf::Mouse::setPosition(sf::Vector2i(pos.x, pos.y), *relativeTo);
+}
+
+extern "C" sfVector2i sfMouse_getPositionRenderWindow(const sf::RenderWindow *relativeTo) {
+    sf::Vector2i vec2 = sf::Mouse::getPosition(*relativeTo);
+    return {vec2.x, vec2.y};
+}
+
+extern "C" sfVector2i sfTouch_getPositionRenderWindow(unsigned int finger, const sf::RenderWindow *relativeTo) {
+    sf::Vector2i pos = sf::Touch::getPosition(finger, *relativeTo);
+    return {pos.x, pos.y};
+}