+++ CSFML/src/System/SfStdVector.cpp
@@ -1,4 +1,3 @@
-#include <SFML/Window/VideoMode.hpp>
 #include <string>
 #include <vector>
 
@@ -11,13 +10,5 @@ extern "C" std::size_t sfStdStringVector_getLength(const std::vector<std::string
 }
 
 extern "C" const std::string *sfStdStringVector_index(const std::vector<std::string> *vec, std::size_t index) {
-    return &(*vec)[index];
-}
-
-extern "C" std::size_t sfVideoModeVector_getLength(const std::vector<sf::VideoMode> *vec) {
-    return vec->size();
-}
-
-extern "C" const sf::VideoMode *sfVideoModeVector_index(const std::vector<sf::VideoMode> *vec, std::size_t index) {
     return &(*vec)[index];
 }