+++ CSFML/src/Window/Mouse.cpp
@@ -1,5 +1,4 @@
 #include "System/Vector2.h"
-#include <SFML/Graphics/RenderWindow.hpp>
 #include <SFML/Window/Mouse.hpp>
 #include <SFML/Window/Window.hpp>
 
@@ -17,19 +16,10 @@ extern "C" sfVector2i sfMouse_getPositionRelativeTo(const sf::Window *relativeTo
     return {pos.x, pos.y};
 }
 
-extern "C" sfVector2i sfMouse_getPositionRenderWindow(const sf::RenderWindow *relativeTo) {
-    sf::Vector2i vec2 = sf::Mouse::getPosition(*relativeTo);
-    return {vec2.x, vec2.y};
-}
-
 extern "C" void sfMouse_setPosition(sfVector2i position) {
     sf::Mouse::setPosition(sf::Vector2i(position.x, position.y));
 }
 
 extern "C" void sfMouse_setPositionRelativeTo(sfVector2i pos, const sf::Window *relativeTo) {
-    sf::Mouse::setPosition(sf::Vector2i(pos.x, pos.y), *relativeTo);
-}
-
-extern "C" void sfMouse_setPositionRenderWindow(sfVector2i pos, const sf::RenderWindow *relativeTo) {
     sf::Mouse::setPosition(sf::Vector2i(pos.x, pos.y), *relativeTo);
 }