+++ CSFML/src/Window/Touch.cpp
@@ -1,5 +1,4 @@
 #include "System/Vector2.h"
-#include <SFML/Graphics/RenderWindow.hpp>
 #include <SFML/Window/Touch.hpp>
 
 extern "C" bool sfTouch_isDown(unsigned int finger) {
@@ -12,11 +11,6 @@ extern "C" sfVector2i sfTouch_getPosition(unsigned int finger) {
 }
 
 extern "C" sfVector2i sfTouch_getPositionRelativeTo(unsigned int finger, const sf::Window *relativeTo) {
-    sf::Vector2i pos = sf::Touch::getPosition(finger, *relativeTo);
-    return {pos.x, pos.y};
-}
-
-extern "C" sfVector2i sfTouch_getPositionRenderWindow(unsigned int finger, const sf::RenderWindow *relativeTo) {
     sf::Vector2i pos = sf::Touch::getPosition(finger, *relativeTo);
     return {pos.x, pos.y};
 }